requirements.txt
absl-py==0.9.0
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
astunparse==1.6.3
Brlapi==0.6.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
configparser==5.0.0
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
future==0.18.2
gast==0.3.3
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.56.4
grpcio==1.27.2
h5py==2.10.0
httplib2==0.21.0
idna==2.9
joblib==0.14.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.0
matplotlib==3.1.3
mysql-connector-python==8.0.31
netifaces==0.10.4
numpy==1.18.1
oauth==1.0.1
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.45.1
opt-einsum==3.1.0
pandas==1.0.1
pathlib2==2.3.5
pbr==5.10.0
pefile==2019.4.18
pexpect==4.2.1
Pillow==9.3.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.26.1
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.1.2
pyparsing==2.4.6
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.3
pyxdg==0.25
PyYAML==5.3
reportlab==3.4.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
shutils==0.1.0
simplejson==3.13.2
six==1.14.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==2.3.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
testresources==2.0.1
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
Werkzeug==1.0.0
wrapt==1.12.0
xkit==0.0.0
zope.interface==4.3.2
